Fix multiple decoration for paramSingle
This fixes the multiple decoration logic for paramSingle where there
are more than one scopes that contain decorators for a type. The lookup
logic only looked up one-level for the decorator. For example, if Scope S
has a decorator and its parent Sp has a decorator for a type, it never
executed Sp's decorator.

This caused the order of the Scopes in which a function that used such type
was Invoked from to change the execution result.

This PR addresses the issue by more diligently looking up the types by
traversing for more decorators until there is no more decorators found
in the path to the root Scope.

Also added several tests that cover this case.