Fix issue with multiple decoration passes in same Scope
This fixes issue when there is a decorator that depends on a type
that is decorated by another decorator that lives in the same
Scope.

Previously, there is a "circuit-breaker" for the parameter building
logic specifically for the case when the decorator depends on the same
type as its output type; this basically causes the decorator.Call and
paramSingle.Build to call each other, ending up in an infinite recursion.

To avoid that, it was skipping over the current Scope when looking for
decorators for its parameters if it was already in the context of a decorator.

However, that logic is incorrect as it is possible for the decorator can still
depend on *other* types that is decorated within the same Scope.

This fixes the issue by keeping track of the decorator's state in 3 modes -
ready, on stack, and called. If a decorator is on stack, that means that it is
still in the middle of being executed, so when parameter builder finds that
decorator, it knows to skip over the current Scope to look for any decorators
from parent Scopes. Otherwise, it can safely look for decorators from the same
Scope since we won't end up in an infinite recursion.

This fixes #328.