Error out when trying to decorate a value group with single value
When a value group is being decorated with a single value, it currently
panics because decorate tries to take element of a non-slice value.

While it's possible to allow this technically, after some consideration,
this should just not be allowed because it can be confusing to distinguish
between the two cases. For example, when providing a slice of slices as a
value group, it's unclear whether or not the target type is the slice of types,
or the slice of slices of types.

In addition, if there is a good way to support it in the future, it is
possible to add that later.

Ref GO-1261