Fix issue with position of value group fields in paramObject and add testcases
When declaring a `dig.In` structure with a soft value group, depending
on the positions of the fields, a value could or could not be provided
to the group.

In the next example, before this change, if the provided constructors
are called, the `Strings` slice would be empty, but if we declared
`Ints` first, the Strings slice would be `{"c", "d", "e"}`

```
type params struct {
    dig.In

    Strings []string `group:"strings,soft"`
    Ints    []int    `group:"ints"`
}

type result struct {
    dig.Out

    String string `group:"strings"`
    Int    int    `group:"ints"`
}

parent.Provide(func() result { return result{String: "c", Int: 10} })
parent.Provide(func() result { return result{String: "d", Int: 11} })
parent.Provide(func() result { return result{String: "e", Int: 12} })
```

This commits corrects this behavior, so that if a constructor for a
field in the structure produces a value for a soft value group it will
be provided to it no matter the position the value group was declared.

This also adds test cases to soft value groups using decorators.