Fix false positives in cycle detection with child scopes.
Currently, cycle detection works by assigning each constructor node
an "order"/index for the scope it was given in. If an edge exists
between an index and itself, a cycle is detected.

When we create a subscope, we copy all of the parent's nodes
to the child scope, but we do not copy their orders.
This causes all root scope constructors to effectively have
order 0 in the child scope (zero value for a mapping to int).
This causes false positives when doing cycle detection
as an edge from order 0 to order 0 gets detected.

This PR fixes this bug by copying the order for constructor nodes
from parent to child scope when a child scope is created.

The added test case fails before this PR, passes with it.

Ref: https://github.com/uber-go/dig/issues/397